+++ Tests/SentryTests/Integrations/Performance/CoreData/SentryCoreDataMiddleware+Extension.swift
@@ -2,12 +2,20 @@ import Foundation
 
 extension SentryCoreDataMiddleware {
     
-    func fetchManagedObjectContext<T>(_ context: NSManagedObjectContext, request: NSFetchRequest<T>, originalImp: (NSFetchRequest<T>, NSErrorPointer) -> [T]?) throws -> [Any] {
+    func fetchManagedObjectContext<T>(_ context: NSManagedObjectContext, request: NSFetchRequest<T>, isErrorNil: Bool = false, originalImp: (NSFetchRequest<T>, NSErrorPointer) -> [T]?) throws -> [Any] {
         
         var error: NSError?
+        var result: [Any]
         
-        let result = __managedObjectContext(context, execute: request as! NSFetchRequest<NSFetchRequestResult>, error: &error) { fetchRequest, errorOut in
-            return originalImp(fetchRequest as! NSFetchRequest<T>, errorOut)
+        if isErrorNil {
+            result = __managedObjectContext(context, execute: request as! NSFetchRequest<NSFetchRequestResult>, error: nil) { fetchRequest, errorOut in
+                return originalImp(fetchRequest as! NSFetchRequest<T>, errorOut)
+            }
+            
+        } else {
+            result = __managedObjectContext(context, execute: request as! NSFetchRequest<NSFetchRequestResult>, error: &error) { fetchRequest, errorOut in
+                return originalImp(fetchRequest as! NSFetchRequest<T>, errorOut)
+            }
         }
         
         if let er = error {