+++ Tests/SentryTests/Integrations/Performance/CoreData/SentryCoreDataTracker+Test.h
@@ -1,0 +1,15 @@
+#import "SentryCoreDataTracker.h"
+#import <Foundation/Foundation.h>
+#import <XCTest/XCTest.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface
+SentryCoreDataTracker (Test)
+
+- (BOOL)saveManagedObjectContextWithNilError:(NSManagedObjectContext *)context
+                                 originalImp:(BOOL(NS_NOESCAPE ^)(NSError **))original;
+
+@end
+
+NS_ASSUME_NONNULL_END