+++ Tests/SentryTests/Integrations/Performance/CoreData/SentryCoreDataTracker+Test.m
@@ -1,0 +1,13 @@
+#import "SentryCoreDataTracker+Test.h"
+#import <Foundation/Foundation.h>
+
+@implementation
+SentryCoreDataTracker (Test)
+
+- (BOOL)saveManagedObjectContextWithNilError:(NSManagedObjectContext *)context
+                                 originalImp:(BOOL(NS_NOESCAPE ^)(NSError **))original
+{
+    return [self managedObjectContext:context save:nil originalImp:original];
+}
+
+@end